+++ cmd/raster/main.go
@@ -15,7 +15,7 @@ import (
 	"github.com/martin-k-m/raster/internal/value"
 )
 
-const version = "0.27.0"
+const version = "0.28.0"
 
 func main() {
 	args := os.Args[1:]